requirements.txt
# =============================================================================
# AI Mustache Generator — Backend Dependencies
# Python 3.11+
# =============================================================================

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# AI / Image Processing
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# Supabase
supabase==2.4.6

# Rate Limiting
slowapi==0.1.9

# Config & Utilities
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
